LayoutTests/http/tests/security/mixedContent/insecure-worker-global-scope.html

<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body onload="runTest()">
<script>
    description("Tests that mixed content check blocks imported insecure script in a global worker scope.");
    jsTestIsAsync = true;

    if (document.location.protocol !== "https:")
        document.location.href = "https://127.0.0.1:8443/security/mixedContent/insecure-worker-global-scope.html";

    if (window.testRunner)
        testRunner.dumpAsText();

    const numberOfTestsToRun = 2;
    let numberOfTestsDone = 0;
    let interval = setInterval(function () {
        if (numberOfTestsDone === numberOfTestsToRun) {
            clearInterval(interval);
            finishJSTest();
        }
    }, 50);
    function runTest() {
        let worker = new Worker("https://127.0.0.1:8443/security/mixedContent/resources/worker-importing-secure-script.js");
        worker.addEventListener('message', function (message) {
            testPassed("Imported, secure worker script is running and sent a message: '" + message.data + "'");
            numberOfTestsDone++;
        });

        worker = new Worker("https://127.0.0.1:8443/security/mixedContent/resources/worker-importing-insecure-script.js");
        worker.addEventListener('message', function (message) {
            if (message.data === "ERROR")
                testPassed("Imported, insecure worker script threw an exception.");
            else
                testFailed("Imported, insecure worker script is running and sent a message: '" + message.data + "'");
            numberOfTestsDone++;
        });
    }
</script>
</body>
</html>
