LayoutTests/http/tests/security/mixedContent/insecure-worker-global-scope-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/insecure-worker-global-scope.html was not allowed to display insecure content from http://127.0.0.1:8000/security/mixedContent/resources/worker-sending-message.js.

Tests that mixed content check blocks imported insecure script in a global worker scope.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Imported, secure worker script is running and sent a message: 'Running.'
PASS Imported, insecure worker script threw an exception.
PASS successfullyParsed is true

TEST COMPLETE
